requirements.txt
langchain
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
ollama
python-dotenv
fastapi
uvicorn
pypdf
docx2txt